requirements.txt
altgraph==0.17.3
anyio==3.6.2
appdirs==1.4.4
arabic-reshaper==3.0.0
aspose-pdf==23.4.0
aspose-words==23.1.0
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
blackbox==0.7.1
bleach==6.0.0
blinker==1.6.2
bottle==0.12.23
bottle-websocket==0.2.9
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
clr-loader==0.2.6
colorama==0.4.6
coloredlogs==15.0.1
configs==3.0.3
contourpy==1.0.7
convertapi==1.5.0
crypto==1.4.1
cryptography==40.0.2
customtkinter==5.2.0
cycler==0.10.0
Cython==0.29.34
darkdetect==0.8.0
deprecation==2.1.0
docutils==0.20.1
Eel==0.15.1
engineering-notation==0.8.0
exceptiongroup==1.1.2
filelock==3.11.0
firebase-admin==6.1.0
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.38.0
fpdf==1.7.2
fsspec==2023.4.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gradio_client==0.1.3
greenlet==2.0.1
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
huggingface-hub==0.13.4
humanfriendly==10.0
idna==2.10
image2pdf==1.0.0
imageio==2.24.0
img2pdf==0.4.4
importlib-metadata==6.8.0
IronPdf==2023.8.6
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsons==1.6.3
keyboard==0.13.5
keyring==24.2.0
kiwisolver==1.4.4
lapx==0.5.4
llvmlite==0.40.0
lxml==4.9.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mbox==0.1.11
mdurl==0.1.2
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.5
Naked==0.1.32
networkx==3.1
ntplib==0.4.0
Nuitka==1.5.7
numba==0.57.0
numpy==1.23.5
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
ordered-set==4.1.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pdf2image==1.16.3
pefile==2022.5.30
pikepdf==6.2.8.post1
Pillow==9.4.0
pkginfo==1.9.6
poppler-utils==0.1.0
printfactory==0.1.1
proto-plus==1.22.2
protobuf==4.23.0
psutil==5.9.4
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.4
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.16.1
pygubu==0.31
pygubu-designer==0.36
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2022.14
PyJWT==2.7.0
PyMatting==1.1.8
PyMsgBox==1.0.9
PyMuPDF==1.21.1
pyparsing==2.4.7
PyPDF2==3.0.1
pypdftk==0.5
pyperclip==1.8.2
pyproject_hooks==1.0.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyreadline3==3.4.1
PyRect==0.2.0
Pyrogram==2.0.97
PyScreeze==0.1.28
PySocks==1.7.1
pyTelegramBotAPI==4.9.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.0
pythonnet==3.0.1
pytweening==1.0.4
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==40.0
reportlab==3.6.12
requests==2.25.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.5.2
roboflow==1.1.2
rsa==4.9
scipy==1.10.0
screeninfo==0.8.1
seaborn==0.12.2
selenium==4.10.0
sentry-sdk==1.19.1
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
supervision==0.13.0
sympy==1.11.1
TgCrypto==1.2.5
thop==0.1.1.post2209072238
tk-tools==0.16.0
tkinterdnd2==0.3.0
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
ttkbootstrap==1.10.1
twine==4.0.2
typing_extensions==4.4.0
typish==1.9.3
ultralytics==8.0.151
ultralytics-yolo==0.0.1
undetected-chromedriver==3.5.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.6
webdriver-manager==4.0.0
webencodings==0.5.1
websockets==11.0.2
Werkzeug==2.3.7
wget==3.2
whichcraft==0.6.1
winaudio==1.0.6
wsproto==1.2.0
zipp==3.16.2
zope.event==4.6
zope.interface==5.5.2
zstandard==0.21.0
